control-server: add ThreadedCommandRunner to control-server
With `control_connection_start_as_thread` it is easy to run
other control commands in a separate thread.
